requirements.txt
numpy
scipy
tensorflow-gpu
Pillow
matplotlib
tqdm